fix(client): Update ProjectProgressService to support JSON files
- Remove unused domain service (moved logic to infrastructure)
- Update file counting to accept both .md and .json extensions
- Fix context/ directory scanning to properly count User Stories (JSON)
- Improve code organization and documentation

This fixes workflow blocking at Step 5 (USER_STORIES_MASTER) where
the system expected only .md files in context/ directory.
