+++ src/client/lib/features/project_shell/infrastructure/services/project_progress_service.dart
@@ -8,59 +8,91 @@ import '../../domain/entities/project_progress.dart';
 import '../../domain/models/project_phase.dart';
 
 /// Service for managing project progress persistence
-///
-/// This service handles creation and updates of `.softarchitect/status.json`
-/// file tracking project completion state.
+/// Merged version: handles both real filesystem scanning and mock projects.
 class ProjectProgressService {
-  /// Path to the hidden folder containing progress data
   static const String _progressFolderName = '.softarchitect';
-
-  /// Name of the status file
   static const String _statusFileName = 'status.json';
 
-  /// Total number of documents expected across all phases
   static int get _totalDocuments =>
       ProjectPhase.all.fold(0, (sum, phase) => sum + phase.fileCount);
 
-  /// Calculate current progress based on document count in context folder
-  ///
-  /// Scans the [projectRoot]/context directory recursively and counts
-  /// all .md files excluding README.md and untitled documents.
-  ///
-  /// Returns [ProjectProgress] with updated metrics.
-  /// Throws [FileSystemException] if projectRoot doesn't exist.
-  static Future<ProjectProgress> calculateProgress(String projectRoot) async {
-    final projectDir = Directory(projectRoot);
+  // --- MÉTODOS DE LA VERSIÓN 2 (Comprobaciones dinámicas y Mocks) ---
 
-    if (!projectDir.existsSync()) {
-      throw FileSystemException('Project root does not exist', projectRoot);
+  /// Checks if a project path is a mock/guide project.
+  static bool isMockProject(String projectPath) =>
+      projectPath.startsWith('mock://');
+
+  /// Calculates the number of documents created in a project.
+  static Future<int> calculateDocumentsCreated(String projectPath) async {
+    if (isMockProject(projectPath)) {
+      return 12;
     }
 
-    // Count markdown files in context/ folder
-    final contextPath = p.join(projectRoot, 'context');
-    final contextDir = Directory(contextPath);
+    try {
+      final contextDir = Directory(p.join(projectPath, 'context'));
+      if (!await contextDir.exists()) {
+        return 0;
+      }
 
-    var documentCount = 0;
-
-    if (contextDir.existsSync()) {
+      var count = 0;
       await for (final entity in contextDir.list(
         recursive: true,
         followLinks: false,
       )) {
-        if (entity is File && entity.path.endsWith('.md')) {
-          final filename = p.basename(entity.path).toLowerCase();
-          // Exclude README and untitled documents
-          if (!filename.contains('readme') && !filename.contains('untitled')) {
-            documentCount++;
+        if (entity is File) {
+          final path = entity.path.toLowerCase();
+          // 🎯 FIX: Ahora permitimos tanto .md como .json para que el paso 5 cuente
+          if (path.endsWith('.md') || path.endsWith('.json')) {
+            final filename = p.basename(path);
+            if (!filename.contains('readme') &&
+                !filename.contains('untitled')) {
+              count++;
+            }
           }
         }
       }
+      return count;
+    } catch (e) {
+      return 0;
+    }
+  }
+
+  /// Gets the current phase name for display.
+  static String getCurrentPhase(int documentsCreated, String projectPath) {
+    if (isMockProject(projectPath)) {
+      return 'Guía Completa';
     }
 
-    // Determine current phase
-    final currentPhase = _determinePhase(documentCount);
+    var accumulated = 0;
+    for (final phase in ProjectPhase.all) {
+      accumulated += phase.fileCount;
+      if (documentsCreated < accumulated) {
+        return phase.name;
+      }
+    }
+    return ProjectPhase.all.last.name;
+  }
 
-    // Calculate completion percentage
+  // --- MÉTODOS DE LA VERSIÓN 1 (Persistencia en status.json) ---
+
+  /// Calculate current progress based on document count in context folder
+  static Future<ProjectProgress> calculateProgress(String projectRoot) async {
+    if (isMockProject(projectRoot)) {
+      return ProjectProgress(
+        documentosCreados: 12,
+        faseActual: 'Guía Completa',
+        porcentajeCompletado: 50,
+        lastUpdated: DateTime.now(),
+      );
+    }
+
+    final projectDir = Directory(projectRoot);
+    if (!projectDir.existsSync()) {
+      throw FileSystemException('Project root does not exist', projectRoot);
+    }
+
+    final documentCount = await calculateDocumentsCreated(projectRoot);
+    final currentPhase = getCurrentPhase(documentCount, projectRoot);
     final percentage = _calculatePercentage(documentCount);
 
     return ProjectProgress(
@@ -72,22 +104,19 @@ class ProjectProgressService {
   }
 
   /// Persists progress data to `.softarchitect/status.json`
-  ///
-  /// Creates the hidden folder if it doesn't exist.
-  /// Overwrites existing status file with new data.
-  ///
-  /// Throws [FileSystemException] on write errors.
   static Future<void> saveProgress({
     required String projectRoot,
     required ProjectProgress progress,
   }) async {
-    // Create .softarchitect folder if doesn't exist
+    if (isMockProject(projectRoot)) {
+      return;
+    }
+
     final progressDir = Directory(p.join(projectRoot, _progressFolderName));
     if (!progressDir.existsSync()) {
       await progressDir.create(recursive: true);
     }
 
-    // Write status.json
     final statusFile = File(p.join(progressDir.path, _statusFileName));
     final jsonString = const JsonEncoder.withIndent(
       '  ',
@@ -97,12 +126,11 @@ class ProjectProgressService {
   }
 
   /// Loads progress data from `.softarchitect/status.json`
-  ///
-  /// Returns [ProjectProgress] if file exists and is valid JSON.
-  /// Returns null if file doesn't exist.
-  ///
-  /// Returns null if file doesn't exist or JSON is malformed.
   static Future<ProjectProgress?> loadProgress(String projectRoot) async {
+    if (isMockProject(projectRoot)) {
+      return null;
+    }
+
     final statusFile = File(
       p.join(projectRoot, _progressFolderName, _statusFileName),
     );
@@ -122,34 +150,15 @@ class ProjectProgressService {
   }
 
   /// Updates progress after a new document is saved
-  ///
-  /// Recalculates metrics and persists to disk.
-  /// This should be called after any document save operation.
-  ///
-  /// Example:
-  /// ```dart
-  /// await FileSystemService.saveDocument(...);
-  /// await ProjectProgressService.updateAfterDocumentSave(projectRoot);
-  /// ```
   static Future<ProjectProgress> updateAfterDocumentSave(
     String projectRoot,
   ) async {
+    if (isMockProject(projectRoot)) {
+      return calculateProgress(projectRoot);
+    }
     final progress = await calculateProgress(projectRoot);
     await saveProgress(projectRoot: projectRoot, progress: progress);
     return progress;
-  }
-
-  // Private helper methods
-
-  static String _determinePhase(int documentCount) {
-    var accumulator = 0;
-    for (final phase in ProjectPhase.all) {
-      accumulator += phase.fileCount;
-      if (documentCount < accumulator) {
-        return phase.name;
-      }
-    }
-    return 'Proyecto Completado';
   }
 
   static double _calculatePercentage(int documentCount) {
